Expender POC: test if the MMU faults from an unmapped pte
Let us do it entirely at EL1 and here is what I went down experimenting
a bit. I assumed (as per-notes) we have three sets of addresses to manage
(bit of ASCII diagramming):

|V0---V1--------V3|    VA moving
|P0---P1|              PA_expender, fixed
|F0---F1--------F3|    PA_FPGA_full_range, moving
		       F0, F1 – phys addr exposed at a time

We reserve a great area of VA being the size of the PA_FPGA_full_range.
If the address is currently mapped is being served, if not MMU aborts and
in the handler we map/release the subset of the addresses and steer
the expender at the same time so that it exposes what is being requested.
Probably some details aren’t yet worked out but to me it should be doing
the job sought.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>