Extender POC: algorithm design
Let us assume we have three sets of addresses to manage: virt and phys
addresses of the Operation System (Linux kernel with rootfs) denoted as
V and P respectively, and the phys address of the FPGA fabric denoted as
F as shown in the ASCII diagramming below.

Assume then we reserve a great area of V of the size of
the PA_FPGA_full_range. Linux kernel (or rootfs) uses any subset of V from
that area.

Case #1:
If V used falls into V0-V1 mapped to P0-P1, that in turn is mapped to F0-F1,
F0-F1 from FPGA can be accessed (figure below).

|--V0-V1-------| VA moving, V0-V1 mapped to P0-P1
|---P0-P1------| PA_expender, fixed
|--F0-F1-------| PA_FPGA_full_range, moving
                 F0-F1 exposed from FPGA

Case #2:
If EL1/0 uses the V from V2-V3 not mapped then the MMU aborts.
In the handler of this abort we map V2-V3 to P0-P1 while releasing V0-V1.
The expender at the same time is steered so that it exposes F2-F3 moving
away from F0-F1.

|--V0-V1--V2-V3-| VA moving,V1-V2 unampped, V2-V3 mapped
|---P0-P1-------| PA_expender, fixed
|--F0-F1--F2-F3-| PA_FPGA_full_range, moving
                  Exposure of F0-F1 moved to F1-F2

and so on.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>